requirements.txt
asgiref==3.7.2
cffi==1.16.0
cryptography==41.0.7
Django==4.2.9
django-crispy-forms==1.12.0
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.14.0
MarkupSafe==2.1.4
pycparser==2.21
pyOpenSSL==23.3.0
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
Werkzeug==3.0.1
